reorganize file locations
