+++ lib/commit_lister/git_api/lister.rb
