+++ lib/commit_lister/git_cli/commit_constants.rb
