+++ lib/commit_lister/git_cli/git_wrapper.rb
